code fix old code handling integers. Note this causes overflows.
git-svn-id: https://svn.science.uu.nl/repos/project.fisme.java/DWOJAppletClient/branches/working@10757 58dd2db4-0f11-0410-9df4-da15f00f50bd

+++ src/main/java/fi/dwo/dwojapplet/domain/DwoHelper.java
@@ -484,7 +484,7 @@ public final class DwoHelper {
         currentUser = aCurrentUser;
         if(aCurrentUser != null){
             try {
-                currentFacadeUser = (User) PersistenceFacade.instance().get(aCurrentUser.getUserID(), User.class);
+                currentFacadeUser = (User) PersistenceFacade.instance().get((int) (long) aCurrentUser.getUserID(), User.class);
             } catch (PersistenceException ex) {
                 LOG.log(Level.SEVERE, null, ex);
             }
